test(e2e): un-skip PDF export tests; delete /test/editor mock spec
The "needs puppeteer Chromium in CI" skip on export-pdf-dashboard and
export-pdf-editor is outdated: PDF export now uses the browser's native
window.print() flow (src/lib/pdf/print-export.ts), not server-side puppeteer.

- Rewrite the dashboard test to neuter window.print and assert on the
  rendered popup (title matches doc, .content wrapper present).
- Delete export-pdf-editor.spec.ts entirely. All 4 tests used the
  /test/editor mock page (CLAUDE.md violation) and tested the same
  export button that pdf-visual-regression.spec.ts already verifies
  from a real seeded document with image-diff per fixture.
- Extract neuterPopupPrint / settlePopup into e2e/helpers/pdf-export.ts
  so the dashboard spec and the visual regression spec share one impl.

Net: 4 tests un-skipped (1 newly running in CI, 3 redundant deleted).
